+++ camd/tests/test_analysis.py
@@ -3,10 +3,11 @@
 import unittest
 import os
 import pandas as pd
+import pickle
 from monty.serialization import loadfn
 from monty.tempfile import ScratchDir
 from camd import CAMD_TEST_FILES
-from camd.analysis import StabilityAnalyzer, AnalyzeStructures
+from camd.analysis import StabilityAnalyzer, AnalyzeStructures, GenericATFAnalyzer
 from camd.utils.data import filter_dataframe_by_composition
 
 
@@ -65,6 +66,22 @@ class StructureAnalyzerTest(unittest.TestCase):
 
         self.assertEqual(analyzer.analyze_vaspqmpy_jobs(jobs, against_icsd=True, use_energies=True),
                          [True, True, True])
+
+
+class GenericATFAnalyzerTest(unittest.TestCase):
+    def test_analyze(self):
+        seed_size = 10
+        exploration_df = pd.read_csv(os.path.join(CAMD_TEST_FILES, "test_df_ATF.csv"))
+        record = pickle.load(open(os.path.join(CAMD_TEST_FILES, "seed_data_ATF.pickle"), "rb"))
+        analyzer = GenericATFAnalyzer(exploration_df, percentile=0.01)
+        seed_data = pd.DataFrame(record[:seed_size])
+        new_experimental_results = pd.DataFrame(record[seed_size:])
+        summary, _ = analyzer.analyze(new_experimental_results, seed_data)
+
+        self.assertEqual(summary['deALM'].to_list()[0].shape[0], record.shape[0]-seed_size)
+        self.assertEqual(summary['anyALM'].to_list()[0].shape[0], record.shape[0]-seed_size)
+        self.assertEqual(summary['allALM'].to_list()[0].shape[0], record.shape[0]-seed_size)
+        self.assertEqual(summary['simALM'].to_list()[0].shape[0], record.shape[0]-seed_size)
 
 
 if __name__ == '__main__':