+++ camd/tests/test_domain.py
@@ -1,7 +1,6 @@
 import unittest
 import os
-from pymatgen import Structure
-from pymatgen.symmetry.analyzer import SpacegroupAnalyzer
+from pymatgen.core.structure import Structure
 
 from camd.domain import StructureDomain, get_structures_from_protosearch, heuristic_setup